requirements.txt
pandas==0.25.1
numpy==1.16.4
matplotlib==3.1.1
joblib==0.14.1
requests==2.22.0
beautifulsoup4==4.8.1
python-dateutil==2.8.1